Refactor code
Refactor code to improve readability and enhance comprehension. Add comments to clarify functionality and simplify code maintenance. Update certain sections of the code to adhere to better practices and ensure greater flexibility.

+++ Data/Resources.cs
@@ -3,11 +3,17 @@ using System.IO;
 
 namespace Axios.Data
 {
+    /// <summary>
+    /// A class for managing application resources such as garbage collector and cache.
+    /// </summary>
     internal class Resources
     {
-        public static readonly string TEMP_FOLDER_PATH = Path.GetTempPath() + "AxiosTemp";
-        public static readonly string CACHE_FILE_PATH = Path.GetTempPath() + "AxiosTemp\\StationCache.json";
+        public static readonly string TempFolderPath = Path.GetTempPath() + "AxiosTemp";
+        public static readonly string CacheFilePath = Path.GetTempPath() + "AxiosTemp\\StationCache.json";
 
+        /// <summary>
+        /// Forces garbage collection and waits for finalizers to finish, repeatedly collecting until no memory is left.
+        /// </summary>
         public static void EnforceClean()
         {
             int collectionCount = 0;
@@ -19,34 +25,32 @@ namespace Axios.Data
             } while (GC.GetTotalMemory(false) != 0 && collectionCount < 10);
         }
 
+        /// <summary>
+        /// Initializes the temporary directory if it does not exist.
+        /// </summary>
         public static void InitializeTempDir()
         {
-            if (!Directory.Exists(TEMP_FOLDER_PATH))
-            {
-                Directory.CreateDirectory(TEMP_FOLDER_PATH);
-            }
+            if (!Directory.Exists(TempFolderPath)) { Directory.CreateDirectory(TempFolderPath); }
         }
 
+        /// <summary>
+        /// Clears the temporary directory by deleting all files.
+        /// </summary>
+        /// <param name="deleteJsonCache">Whether or not to delete the JSON cache file.</param>
+        /// <exception cref="Exception">Thrown when failed to clear the temp directory.</exception>
         public static void ClearTempDir(bool deleteJsonCache = false)
         {
             try
             {
-                if (Directory.Exists(TEMP_FOLDER_PATH))
+                if (!Directory.Exists(TempFolderPath)) { return; }
+
+                foreach (string file in Directory.GetFiles(TempFolderPath))
                 {
-                    foreach (string file in Directory.GetFiles(TEMP_FOLDER_PATH))
-                    {
-                        if (Path.GetExtension(file) != ".json")
-                        {
-                            File.Delete(file);
-                        }
-                        else
-                        {
-                            if (deleteJsonCache) { File.Delete(file); }
-                        }
-                    }
+                    if (Path.GetExtension(file) != ".json") { File.Delete(file); }
+                    else { if (deleteJsonCache) { File.Delete(file); } }
                 }
             }
-            catch (Exception) { return; }
+            catch (Exception e) { throw new Exception("Failed to clear temp directory.", e); }
         }
     }
 }
